+++ pennylane_lightning/core/src/gates/GateOperation.hpp
@@ -69,6 +69,7 @@ enum class GateOperation : uint32_t {
     /* Multi-qubit gates */
     MultiRZ,
     GlobalPhase,
+    PCPhase,
     /* END (placeholder) */
     END
 };
@@ -104,6 +105,7 @@ enum class ControlledGateOperation : uint32_t {
     /* Multi-qubit gates */
     MultiRZ,
     GlobalPhase,
+    PCPhase,
     /* END (placeholder) */
     END
 };
@@ -182,5 +184,23 @@ enum class ControlledMatrixOperation : uint32_t {
     /* END (placeholder) */
     END
 };
+
+/**
+ * @brief Enum class for sparse matrix operation
+ */
+enum class SparseMatrixOperation : uint32_t {
+    BEGIN = 0,
+    SparseMultiQubitOp = 0,
+    /* END (placeholder) */
+    END
+};
+
+enum class ControlledSparseMatrixOperation : uint32_t {
+    BEGIN = 0,
+    NCSparseMultiQubitOp = 0,
+    /* END (placeholder) */
+    END
+};
+
 /// @endcond
 } // namespace Pennylane::Gates